fix: header building

+++ crates/rpc/src/server/mod.rs
@@ -86,7 +86,7 @@ mod tests {
         // https://github.com/hyperium/tonic/issues/1040#issuecomment-1191832200
         let mut message = Vec::new();
         message.push(0);
-        message.extend_from_slice(&0u64.to_be_bytes());
+        message.extend_from_slice(&0u32.to_be_bytes());
 
         let response = client
             .post(format!("http://{rpc_addr}/rpc.Api/Status"))
